pyproject.toml
[tool.poetry]
name = "Evaluation Sandbox"
version = "0.0.1"
description = ""
authors = ["Ira Rainey"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
openai = "^1.57.4"
azure-ai-evaluation = "^1.1.0"
marshmallow = "3.23.2"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.13.0"
flake8 = "^7.1.1"
pylint = "^3.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
